requirements.txt
# python 3.8.7
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
crcmod==1.7
crypto==1.4.1
ddddocr==1.4.10
ddt==1.7.0
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.2.0
flatbuffers==23.5.26
google==3.0.0
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.6.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==21.3.0
iniconfig==2.0.0
jaraco.classes==3.3.1
keyring==24.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
Naked==0.1.31
nh3==0.2.15
numpy==1.24.4
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parameterized==0.9.0
Pillow==10.1.0
pkginfo==1.9.6
pluggy==1.3.0
protobuf==4.25.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.10.1
pyDes==2.0.1
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyotp==2.9.0
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==5.4.1
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rsa==4.7.2
selenium==4.15.2
shadowsocks==2.8.2
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sympy==1.12
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
twine==4.0.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
wsproto==1.2.0
xlrd==2.0.1
zipp==3.17.0
